Fixed compile warning
Fixed numerous compile warnings
TEST-180 Functional tests: add --vm-name-prefix parameter
TEST-180 Functional tests: add debug printing to nat box provision script
VMS-5524 Change auto-generated certificate algorithm to SHA2 or SHA256
VMS-5561 Sometimes HardwareID can't be calculated

<==== Mercurial data ====>
Branch: prod_3.0.0

+++ unit_tests/nx_network_ut/src/udp_socket_ut.cpp
@@ -65,7 +65,7 @@ void udpSocketTransferTest(int ipVersion, const HostAddress& targetHost)
             ASSERT_EQ(SystemError::noError, code);
             ASSERT_TRUE(ip.address.isIpAddress());
             ASSERT_EQ(receiverEndpoint.toString(), ip.toString());
-            ASSERT_EQ(kTestMessage.size(), size);
+            ASSERT_EQ((size_t)kTestMessage.size(), size);
             sendPromise.set_value();
         });
